Throw a specific exception so we can get stderr
We're having a problem with command_line_tools security-update
where all the plugins get updated, but it says:
Error, Only updated 6 of 6 plugins.

I suspect one of the premium plugins doesn't update something
correctly after updating it's files so I want to catch this error and
ignore it.

Having stderr in the exception lets us check whether we're getting this
particular error when executing the plugin update.